Rename GroupFilterModal step components to reflect their data
